- Moved user context from entire app to just the app layout.
