+++ Website/src/features/app/hooks/useAuth.ts
@@ -1,10 +1,7 @@
 import type { getUserResponse } from "@/api/AuthApi";
 import { createContext, useContext } from "react";
 
-export type AuthContextTypes = getUserResponse & {
-  refreshUser: () => void;
-};
-
+type AuthContextTypes = getUserResponse;
 export const AuthContext = createContext<AuthContextTypes | null>(null);
 
 export function useAuth() {
